+++ cmd/root.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const Version = "1.2.0"
+const Version = "1.3.0"
 
 var rootCmd = &cobra.Command{
 	Use:   "snare",
@@ -26,6 +26,7 @@ func init() {
 	rootCmd.AddCommand(showCmd)
 	rootCmd.AddCommand(diffCmd)
 	rootCmd.AddCommand(replayCmd)
+	rootCmd.AddCommand(importCmd)
 	rootCmd.AddCommand(saveCmd)
 	rootCmd.AddCommand(exportCmd)
 	rootCmd.AddCommand(clearCmd)